+++ app/src/main/java/com/trendyol/design/ui/component/inputfield/Dropdown.kt
@@ -3,6 +3,7 @@ package com.trendyol.design.ui.component.inputfield
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.width
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -19,6 +20,7 @@ import com.trendyol.design.ui.theme.TrendyolTheme
 
 private const val LABEL = "Label"
 private const val SELECT = "Select"
+private const val MIN_WIDTH = 280
 
 @Preview(showBackground = true)
 @ShowkaseComposable(
@@ -30,22 +32,25 @@ private const val SELECT = "Select"
 internal fun Input_Field_3_Dropdown_1_Unfocused() = TrendyolTheme {
     Column {
         Dropdown(
+            modifier = Modifier.width(MIN_WIDTH.dp),
             style = TrendyolDropdownStyle.FloatingLabelOutlined,
             label = LABEL,
             value = "",
-            onClicked = {},
+            onClick = {},
         )
         Spacer(modifier = Modifier.height(8.dp))
         Dropdown(
+            modifier = Modifier.width(MIN_WIDTH.dp),
             style = TrendyolDropdownStyle.Outlined,
             value = "",
-            onClicked = {},
+            onClick = {},
         )
         Spacer(modifier = Modifier.height(8.dp))
         Dropdown(
+            modifier = Modifier.width(MIN_WIDTH.dp),
             style = TrendyolDropdownStyle.Filled,
             value = "",
-            onClicked = {},
+            onClick = {},
         )
     }
 }
@@ -60,22 +65,25 @@ internal fun Input_Field_3_Dropdown_1_Unfocused() = TrendyolTheme {
 internal fun Input_Field_3_Dropdown_2_Typed() = TrendyolTheme {
     Column {
         Dropdown(
+            modifier = Modifier.width(MIN_WIDTH.dp),
             style = TrendyolDropdownStyle.FloatingLabelOutlined,
             label = LABEL,
             value = SELECT,
-            onClicked = {},
+            onClick = {},
         )
         Spacer(modifier = Modifier.height(8.dp))
         Dropdown(
+            modifier = Modifier.width(MIN_WIDTH.dp),
             style = TrendyolDropdownStyle.Outlined,
             value = SELECT,
-            onClicked = {},
+            onClick = {},
         )
         Spacer(modifier = Modifier.height(8.dp))
         Dropdown(
+            modifier = Modifier.width(MIN_WIDTH.dp),
             style = TrendyolDropdownStyle.Filled,
             value = SELECT,
-            onClicked = {},
+            onClick = {},
         )
     }
 }
@@ -91,23 +99,26 @@ internal fun Input_Field_3_Dropdown_3_Disabled() = TrendyolTheme {
     Column {
         val value by remember { mutableStateOf("Disabled") }
         Dropdown(
+            modifier = Modifier.width(MIN_WIDTH.dp),
             style = TrendyolDropdownStyle.FloatingLabelOutlined,
             value = value,
-            onClicked = {},
+            onClick = {},
             enabled = false,
         )
         Spacer(modifier = Modifier.height(8.dp))
         Dropdown(
+            modifier = Modifier.width(MIN_WIDTH.dp),
             style = TrendyolDropdownStyle.Outlined,
             value = value,
-            onClicked = {},
+            onClick = {},
             enabled = false,
         )
         Spacer(modifier = Modifier.height(8.dp))
         Dropdown(
+            modifier = Modifier.width(MIN_WIDTH.dp),
             style = TrendyolDropdownStyle.Filled,
             value = value,
-            onClicked = {},
+            onClick = {},
             enabled = false,
         )
     }
@@ -127,7 +138,7 @@ internal fun Input_Field_3_Dropdown_4_Error() = TrendyolTheme {
             style = TrendyolDropdownStyle.FloatingLabelOutlined,
             label = LABEL,
             value = value,
-            onClicked = {},
+            onClick = {},
             isError = true,
             errorLabel = "Error message goes here",
         )
@@ -135,7 +146,7 @@ internal fun Input_Field_3_Dropdown_4_Error() = TrendyolTheme {
         Dropdown(
             style = TrendyolDropdownStyle.FloatingLabelOutlined,
             value = value,
-            onClicked = {},
+            onClick = {},
             isError = true,
             errorLabel = "Error message goes here",
         )
@@ -143,7 +154,7 @@ internal fun Input_Field_3_Dropdown_4_Error() = TrendyolTheme {
         Dropdown(
             style = TrendyolDropdownStyle.Filled,
             value = value,
-            onClicked = {},
+            onClick = {},
             isError = true,
             errorLabel = "Error message goes here",
         )